requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dateutil==2.8.2
urllib3==2.1.0
alembic==1.13.1
